--- a/Dockerfile.test
+++ b/Dockerfile.test
@@ -3,7 +3,7 @@
 #     docker build -t sqliteviz/test -f Dockerfile.test .
 #
 
-FROM node:12
+FROM node:12.22-buster
 
 RUN set -ex; \
     apt update; \