--- a/lib/sql-js/README.md
+++ b/lib/sql-js/README.md
@@ -35,9 +35,8 @@ SQLite [miscellaneous extensions][3] included:
 2. `transitive_closure` virtual table for
    [Querying Tree Structures in SQLite][11] ([closure.c][8])
 3. `uuid`, `uuid_str` and `uuid_blob` RFC-4122 UUID functions ([uuid.c][9])
-4. `regexp` (hence `REGEXP` operator) and `regexpi` functions ([regexp.c][10])
-5. `percentile` function ([percentile.c][13])
-6. `decimal`, `decimal_cmp`, `decimal_add`, `decimal_sub` and `decimal_mul` functions
+4. `percentile` function ([percentile.c][13])
+5. `decimal`, `decimal_cmp`, `decimal_add`, `decimal_sub` and `decimal_mul` functions
    ([decimal.c][14])
 
 SQLite 3rd party extensions included:
@@ -45,6 +44,8 @@ SQLite 3rd party extensions included:
 1. [pivot_vtab][5] -- a pivot virtual table
 2. `pearson` correlation coefficient function extension from [sqlean][21]
    (which is part of [squib][20])
+3. `regexp` extension from [sqlean][10] (via PCRE) providing: `REGEXP` operator and
+   `regexp_like`, `regexp_substr`, `regexp_capture`, `regexp_replace` functions
 
 To ease the step to have working clone locally, the build is committed into
 the repository.
@@ -91,7 +92,7 @@ described in [this message from SQLite Forum][12]:
 [7]:  https://sqlite.org/src/file/ext/misc/series.c
 [8]:  https://sqlite.org/src/file/ext/misc/closure.c
 [9]:  https://sqlite.org/src/file/ext/misc/uuid.c
-[10]: https://sqlite.org/src/file/ext/misc/regexp.c
+[10]: https://github.com/nalgeon/sqlean/tree/main/src/regexp
 [11]: https://charlesleifer.com/blog/querying-tree-structures-in-sqlite-using-python-and-the-transitive-closure-extension/
 [12]: https://sqlite.org/forum/forumpost/6ad7d4f4bebe5e06?raw
 [13]: https://sqlite.org/src/file/ext/misc/percentile.c