… lib/sql-js/dist/sql-wasm.wasm