Build Sqlean's [regexp][3] extension based on PCRE2. 

It's +42% in build size:

```diff
- sql-wasm.wasm 742158
+ sql-wasm.wasm 1054448
```

And requires [EMULATE_FUNCTION_POINTER_CASTS][1] ([more details][2] on it) which is undesired for performance reasons.
Benchmark doesn't show much slowdown though:

```
master	firefox 115	select	0.15 s ± 1.5%
pr	firefox 115	select	0.16 s ± 1.2%
master	chromium 128	select	0.14 s ± 1.4%
pr	chromium 128	select	0.14 s ± 1.4%
master	firefox 115	import	0.93 s ± 1.8%
pr	firefox 115	import	0.91 s ± 1.5%
master	chromium 128	import	1.26 s ± 1.5%
pr	chromium 128	import	1.26 s ± 1.5%
```

[result-analysis.pdf](https://github.com/user-attachments/files/16739517/result-analysis.pdf)


[1]: https://emscripten.org/docs/tools_reference/settings_reference.html#emulate-function-pointer-casts
[2]: https://blog.pyodide.org/posts/function-pointer-cast-handling/
[3]: https://github.com/nalgeon/sqlean/tree/0.27.0/src/regexp