@saaj 2024-08-25
Close in favour of https://github.com/lana-k/sqliteviz/pull/118.